Move group factory to group module
