+++ apps/server/src/modules/authorization-rules/rules/group.rule.spec.ts
@@ -7,10 +7,10 @@ import {
 	AuthorizationInjectionService,
 } from '@modules/authorization';
 import { Group } from '@modules/group';
+import { groupFactory } from '@modules/group/testing';
 import { Test, TestingModule } from '@nestjs/testing';
 import { Role, SchoolEntity, User } from '@shared/domain/entity';
 import { Permission } from '@shared/domain/interface';
-import { groupFactory } from '@testing/factory/domainobject';
 import { roleFactory } from '@testing/factory/role.factory';
 import { schoolEntityFactory } from '@testing/factory/school-entity.factory';
 import { userFactory } from '@testing/factory/user.factory';
