+++ apps/server/src/modules/group/testing/index.ts
@@ -1,1 +1,2 @@
 export * from './group-entity.factory';
+export * from './group.factory';
