+++ apps/server/src/modules/learnroom/repo/mikro-orm/course.repo.integration.spec.ts
@@ -5,7 +5,7 @@ import { ClassEntity } from '@modules/class/entity';
 import { classEntityFactory } from '@modules/class/entity/testing';
 import { Group } from '@modules/group';
 import { GroupEntity } from '@modules/group/entity';
-import { groupEntityFactory } from '@modules/group/testing';
+import { groupEntityFactory, groupFactory } from '@modules/group/testing';
 import { Test, TestingModule } from '@nestjs/testing';
 import {
 	Course as CourseEntity,
@@ -19,7 +19,6 @@ import { SortOrder } from '@shared/domain/interface';
 import { cleanupCollections } from '@testing/cleanup-collections';
 import { courseFactory as courseEntityFactory } from '@testing/factory/course.factory';
 import { courseGroupFactory as courseGroupEntityFactory } from '@testing/factory/coursegroup.factory';
-import { groupFactory } from '@testing/factory/domainobject';
 import { schoolEntityFactory } from '@testing/factory/school-entity.factory';
 import { userFactory } from '@testing/factory/user.factory';
 import { Course, COURSE_REPO, CourseProps, CourseStatus } from '../../domain';
