+++ apps/server/src/modules/learnroom/service/group-deleted-handler.service.spec.ts
@@ -2,9 +2,9 @@ import { createMock, DeepMocked } from '@golevelup/ts-jest';
 import { MikroORM } from '@mikro-orm/core';
 import { ObjectId } from '@mikro-orm/mongodb';
 import { GroupDeletedEvent } from '@modules/group';
+import { groupFactory } from '@modules/group/testing';
 import { Test, TestingModule } from '@nestjs/testing';
 import { Logger } from '@src/core/logger';
-import { groupFactory } from '@testing/factory/domainobject';
 import { setupEntities } from '@testing/setup-entities';
 import { Course } from '../domain';
 import { courseFactory } from '../testing';
