+++ apps/server/src/modules/video-conference/service/video-conference.service.spec.ts
@@ -2,8 +2,17 @@ import { createMock, DeepMocked } from '@golevelup/ts-jest';
 import { CalendarEventDto, CalendarService } from '@infra/calendar';
 import { ObjectId } from '@mikro-orm/mongodb';
 import { AuthorizationContextBuilder, AuthorizationService } from '@modules/authorization';
+import { BoardNodeAuthorizable, BoardNodeAuthorizableService, BoardNodeService, BoardRoles } from '@modules/board';
+import { VideoConferenceElement } from '@modules/board/domain';
+import { columnBoardFactory, videoConferenceElementFactory } from '@modules/board/testing';
+import { GroupTypes } from '@modules/group';
+import { groupFactory } from '@modules/group/testing';
 import { CourseService } from '@modules/learnroom/service';
 import { LegacySchoolService } from '@modules/legacy-school';
+import { RoomService } from '@modules/room';
+import { RoomMembershipService } from '@modules/room-membership';
+import { roomMembershipFactory } from '@modules/room-membership/testing';
+import { roomFactory } from '@modules/room/testing';
 import { UserService } from '@modules/user';
 import { BadRequestException, ForbiddenException } from '@nestjs/common';
 import { NotFoundException } from '@nestjs/common/exceptions/not-found.exception';
@@ -14,16 +23,7 @@ import { Course, TeamUserEntity } from '@shared/domain/entity';
 import { Permission, RoleName, VideoConferenceScope } from '@shared/domain/interface';
 import { EntityId, SchoolFeature } from '@shared/domain/types';
 import { TeamsRepo, VideoConferenceRepo } from '@shared/repo';
-import { BoardNodeAuthorizable, BoardNodeAuthorizableService, BoardNodeService, BoardRoles } from '@src/modules/board';
-import { VideoConferenceElement } from '@src/modules/board/domain';
-import { columnBoardFactory, videoConferenceElementFactory } from '@src/modules/board/testing';
-import { GroupTypes } from '@src/modules/group';
-import { RoomService } from '@src/modules/room';
-import { RoomMembershipService } from '@src/modules/room-membership';
-import { roomMembershipFactory } from '@src/modules/room-membership/testing';
-import { roomFactory } from '@src/modules/room/testing';
 import { courseFactory } from '@testing/factory/course.factory';
-import { groupFactory } from '@testing/factory/domainobject';
 import { roleFactory } from '@testing/factory/role.factory';
 import { teamFactory } from '@testing/factory/team.factory';
 import { teamUserFactory } from '@testing/factory/teamuser.factory';
