--- apps/server/src/testing/factory/domainobject/groups/index.ts
@@ -1,1 +1,0 @@
-export * from './group.factory';
